New GPU backend based on CuPy (#210)
* adds missing requirements

* adds multi-bunch monitor support

* deletes unneeded version

* minor bug fixes

* not running reduce_histo automatically

* monitors record min max

* updates monitor arguments, adds new plot scripts

* debugs mew monitor args

* new monitor configurations

* mods to support larger number of  particles

* cosmetic improvements

* cuda-shmem-hot-bins.ipynb

* sets up efficient compile mechanism

* updates installation instructions

* Fixes py36 compatibility issue

* updates latest fftw version

* programmatic way to figure out python interpreter

* removes timing directives

* prints pycuda location

* Code clean-up

* fixes typo

* improves compile

* updates gpu main scripts

* CGA class to cupy

* conversion to cupy

* MyGpuarray without subclassing

* stdKernel as function

* minor change

* replaced butils_wrap.stdKernel with cupy.std

* minor fixes to kernel calls

* minor change to dev_my_array

* kernels and fft changes

* scaling to irfft

* elw kernels and fft changes

* fft changes

* cupy.rfft without plan

* remove GPUarray inheritance

* minor change to find cupy path

* cupy_array fix

* cupy_array remake

* cupy_array with unified memory and cleanup

* commented out invalidate_cpu/gpu)

* removed enable_gpucache()

* added None object init handler

* CGA.__init__ changes

* minor changes

* cuda kernels clean-up

* cupy install instructions

* removed pycuda unused imports

* minor change

* ports impedance to new simplified gpu scheme

* ports beam to new simplified gpu scheme

* modifies code to be compatible with cupy

* adds cupy/ numpy functions in bmath dictionary

* ports profile to new simplified gpu scheme

* converts beam_feedback

* same api in physics wrap

* function reposition

* ports rf_params

* cleans up bmath

* updates access to rawmodule

* ports tracker

* gpu code clean-up, ready to test

* debugs example 01, runs correctly

* more debugging, examples 2,3 run correctly

* example 4 works, example 5 needs debugging

* prepares remaining examples to be tested

* debugs until example 7, continue with 8

* minor change

* minor changes to gpu_slice

* transfers to gpu imported objects recursively

* all examples produce correct results

* EX01 supports plotting

* cleaning up

* different cubin name, for multiple GPU archs

* corrects unittests

* minor nested loop optimizations

* bug fixes in cupy backend

* minor fixes

* minor fix

* changes for main testcases

* changes for LHC testcase

* minor fixes

* changes for LHC

* performance improvement in beam_feedback

* reverting commits

* minor LHC fix

* typo corrections

* minor changes to testcases

* cupy v11 release

* changes to rf_params calls

* up to date with latest dev in master

* minor change

* debugging the examples

* tested GPU examples, 3 need debugging

* debugged all example files

* wires inter_const_space to interp

* minor bug fixz

* up to date with latest changes

* fixes failing unittests in bmath

* fixes failing unittests in test_synch_rad

* fixes failing unittests in impedance_sources

* fixes warnings in beam

* fixes file not found errors

* fixes failing tests in tracker

* fixes failing examples

* minor improvements in example files

* check for gpu use

* adding unittests for the GPU parts

* fixes failing cupy fft testcases

* adds cp.convolve unittests

* adds unittests for beam_phase

* name mangling attribute fix

* adds unittests for more physics modules

* switch __device to _device

* finished testing all physics kernels

* allreduce compatible with cupy arrays

* adds unittests for impedance related classes

* tests ind volt track

* adds __init__.py to avoid name collisions

* uses pytest parametrize to reduce code length

* fixes warnings

* adds unittests for impedance with MTW

* workers check

* minor changes

* gpu single precision fix

* cleaning up unused code

* function name changes

* gpu drift minor fix

* reduce_histo cpu transfer

* readme changes

* nvcc from cuda path

* lik simplify

* error corrections

* Disables interactive matplotlib backend

* skips failing unittests, adds FIXME note

* changes np.int (deprecated) to int

* removes python3.6 from windows testing

---------

+++ blond/compile.py
@@ -1,4 +1,3 @@
-
 # Copyright 2016 CERN. This software is distributed under the
 # terms of the GNU General Public Licence version 3 (GPL Version 3),
 # copied verbatim in the file LICENCE.md.
@@ -29,27 +28,31 @@ path = os.path.realpath(__file__)
 basepath = os.sep.join(path.split(os.sep)[:-1])
 
 parser = argparse.ArgumentParser(description='Run python setup_cpp.py to'
-                                 ' compile the cpp routines needed from BLonD')
+                                             ' compile the cpp routines needed from BLonD')
 
 parser.add_argument('-p', '--parallel',
                     default=False, action='store_true',
                     help='Produce Multi-threaded code. Use the environment'
-                    ' variable OMP_NUM_THREADS=xx to control the number of'
-                    ' threads that will be used.'
-                    ' Default: Serial code')
+                         ' variable OMP_NUM_THREADS=xx to control the number of'
+                         ' threads that will be used.'
+                         ' Default: Serial code')
 
 parser.add_argument('-b', '--boost', type=str, nargs='?', const='',
                     help='Use boost library to speedup synchrotron radiation'
-                    ' routines. If the installation path of boost differs'
-                    ' from the default, you have to pass it as an argument.'
-                    ' Default: Boost will not be used')
+                         ' routines. If the installation path of boost differs'
+                         ' from the default, you have to pass it as an argument.'
+                         ' Default: Boost will not be used')
 
 parser.add_argument('-c', '--compiler', type=str, default='g++',
                     help='C++ compiler that will be used to compile the'
-                    ' source files. Default: g++')
+                         ' source files. Default: g++')
 
 parser.add_argument('--with-fftw', action='store_true',
                     help='Use the FFTs from FFTW3.')
+
+parser.add_argument('-gpu', '--gpu', nargs='?', const='discover', default=None,
+                    help='Compile the GPU kernels too.'
+                    'Default: Only compile the C++ library.')
 
 parser.add_argument('--with-fftw-threads', action='store_true',
                     help='Use the multi-threaded FFTs from FFTW3.')
@@ -72,9 +75,11 @@ parser.add_argument('--libs', type=str, default='',
 parser.add_argument('-libname', '--libname', type=str, default=os.path.join(basepath, 'cpp_routines/libblond'),
                     help='The blond library name, without the file extension.')
 
+parser.add_argument('-optimize', '--optimize', action='store_true',
+                    help='Auto optimize the compiled library.')
+
 # Additional libs needed to compile the blond library
 libs = []
-
 
 # EXAMPLE FLAGS: -Ofast -std=c++11 -fopt-info-vec -march=native
 #                -mfma4 -fopenmp -ftree-vectorizer-verbose=1
@@ -97,26 +102,36 @@ cpp_files = [
 ]
 
 
-if (__name__ == "__main__"):
+# Get nvcc from CUDA_PATH
+cuda_path = os.getenv('CUDA_PATH', default='')
+if cuda_path != '':
+    nvcc = cuda_path + '/bin/nvcc'
+else:
+    nvcc = 'nvcc'
+
+nvccflags = [nvcc, '--cubin', '-O3', '--use_fast_math', '-maxrregcount', '32']
+# nvccflags = ['nvcc', '--cubin', '-arch', 'sm_xx', '-O3', '--use_fast_math']
+
+if __name__ == "__main__":
     args = parser.parse_args()
     boost_path = None
     with_fftw = args.with_fftw or args.with_fftw_threads or args.with_fftw_omp or \
         (args.with_fftw_lib is not None) or (args.with_fftw_header is not None)
-    if(args.boost is not None):
-        if(args.boost):
+    if args.boost is not None:
+        if args.boost:
             boost_path = os.path.abspath(args.boost)
         else:
             boost_path = ''
         cflags += ['-I', boost_path, '-DBOOST']
     compiler = args.compiler
 
-    if (args.libs):
+    if args.libs:
         libs = args.libs.split()
 
-    if (args.parallel):
+    if args.parallel:
         cflags += ['-fopenmp', '-DPARALLEL', '-D_GLIBCXX_PARALLEL']
 
-    if (args.flags):
+    if args.flags:
         cflags += args.flags.split()
 
     if with_fftw:
@@ -136,17 +151,48 @@ if (__name__ == "__main__"):
                 cflags += ['-DFFTW3PARALLEL']
                 libs += ['-lfftw3_threads', '-lfftw3f_threads']
 
-    if ('posix' in os.name):
+    if 'posix' in os.name:
         cflags += ['-fPIC']
+        if args.optimize:
+            # Check compiler defined directives
+            # This is compatible with python3.6 - python 3.9
+            # The universal_newlines argument transforms output to text (from binary)
+            ret = subprocess.run([compiler + ' -march=native -dM -E - < /dev/null | egrep "SSE|AVX|FMA"'],
+                                 shell=True, stdout=subprocess.PIPE, universal_newlines=True)
+
+            # If we have an error
+            if ret.returncode != 0:
+                print('Compiler auto-optimization did not work. Error: ', ret.stdout)
+            else:
+                # Format the output list
+                stdout = ret.stdout.replace('#define ', '').replace(
+                    '__ 1', '').replace('__', '').split('\n')
+                # Add the appropriate vectorization flag (not use avx512)
+                if 'AVX2' in stdout:
+                    cflags += ['-mavx2']
+                elif 'AVX' in stdout:
+                    cflags += ['-mavx']
+                elif 'SSE4_2' in stdout or 'SSE4_1' in stdout:
+                    cflags += ['-msse4']
+                elif 'SSE3' in stdout:
+                    cflags += ['-msse3']
+                else:
+                    cflags += ['-msse']
+
+                # Add FMA if supported
+                if 'FMA' in stdout:
+                    cflags += ['-mfma']
+
         root, ext = os.path.splitext(args.libname)
         if not ext:
             ext = '.so'
-        libname = root + ext
-    elif ('win' in sys.platform):
+        libname = os.path.abspath(root + ext)
+
+    elif 'win' in sys.platform:
         root, ext = os.path.splitext(args.libname)
         if not ext:
             ext = '.dll'
-        libname = root + ext
+        libname = os.path.abspath(root + ext)
 
         if hasattr(os, 'add_dll_directory'):
             directory, filename = os.path.split(libname)
@@ -169,23 +215,73 @@ if (__name__ == "__main__"):
         print('FFTW3 Library path: ', args.with_fftw_lib)
         print('FFTW3 Headers path: ', args.with_fftw_header)
     print('C++ Compiler: ', compiler)
+    print('Compiler version: ')
+    subprocess.run([compiler, '--version'])
     print('Compiler flags: ', ' '.join(cflags))
     print('Extra libraries: ', ' '.join(libs))
-    subprocess.call([compiler, '--version'])
+    print('Library name: ', libname)
 
-    try:
-        os.remove(libname)
-    except OSError as e:
-        pass
+    # If it exists already, remove the library before re-compiling
+    if os.path.isfile(libname):
+        try:
+            os.remove(libname)
+        except OSError as e:
+            pass
 
-    subprocess.call(command)
+    print('Compiling:\n', ' '.join(command))
+    ret = subprocess.run(command)
+    if ret.returncode != 0:
+        print('\nThere was a compilation error.')
+    else:
+        try:
+            if ('win' in sys.platform) and hasattr(os, 'add_dll_directory'):
+                libblond = ctypes.CDLL(libname, winmode=0)
+            else:
+                libblond = ctypes.CDLL(libname)
+            print('\nThe blond library has been successfully compiled.')
+        except Exception as e:
+            print('\nCompilation failed.')
+            print(e)
 
-    try:
-        if ('win' in sys.platform) and hasattr(os, 'add_dll_directory'):
-            libblond = ctypes.CDLL(libname, winmode=0)
+    # Compile the GPU library
+    if args.gpu:
+        print('\n'+''.join(['=']*80))
+        print('\nCompiling the CUDA library')
+        if args.gpu == 'discover':
+            print('Discovering the device compute capability..')
+            import cupy as cp
+
+            dev = cp.cuda.Device(0)
+            dev_name = cp.cuda.runtime.getDeviceProperties(dev)['name']
+            comp_capability = dev.compute_capability
+            print('Device name {}'.format(dev_name))
+        elif args.gpu is not None:
+            comp_capability = args.gpu
+
+        print('Compiling the CUDA library for architecture {}.'.format(comp_capability))
+        # Add the -arch required argument
+        nvccflags += ['-arch', 'sm_{}'.format(comp_capability)]
+        libname_double = os.path.join(basepath,
+                                      f'gpu/cuda_kernels/kernels_double_sm_{comp_capability}.cubin')
+        libname_single = os.path.join(basepath,
+                                      f'gpu/cuda_kernels/kernels_single_sm_{comp_capability}.cubin')
+        # we need to get the header files location
+        path = cp.__file__.split('/')[:-1]  # remove __init__.py from path
+        path.extend(['_core', 'include'])
+
+        cupyloc = os.path.join('/'.join(path))
+
+        print('cupy: ', cupyloc)
+
+        command = nvccflags + ['-o', libname_single, '-I'+cupyloc,
+                               os.path.join(basepath, 'gpu/cuda_kernels/kernels_single.cu')]
+        subprocess.call(command)
+
+        command = nvccflags + ['-o', libname_double, '-I'+cupyloc,
+                               os.path.join(basepath, 'gpu/cuda_kernels/kernels_double.cu')]
+        subprocess.call(command)
+
+        if os.path.isfile(libname_single) and os.path.isfile(libname_double):
+            print('The CUDA library has been successfully compiled.')
         else:
-            libblond = ctypes.CDLL(libname)
-        print('\nThe blond library has been successfully compiled.')
-    except Exception as e:
-        print('\nCompilation failed.')
-        print(e)
+            print('The CUDA library compilation failed.')
